refactor(runtime): extract message module from agent_loop
The message-shape helper cluster moves into `agent_loop/message.rs`:

- `ACCUMULATED_TEXT_MAX_BYTES` + `push_accumulated_text` — the bounded
  64-KiB fallback buffer the empty-response guard reads when no text
  arrives alongside `tool_use` blocks; sealed-on-overflow with ASCII
  padding so subsequent appends short-circuit.
- `is_no_reply`, `is_progress_text_leak`, `is_cascade_leak` — model-
  output classifiers (delegates to `silent_response` for two of them).
- `is_soft_error_content`, `is_parameter_error_content` — tool-error
  content classifiers used by the consecutive-failure abort threshold.
- `sanitize_for_memory` (+ the inline `use` of
  `ENVELOPE_LINE_PREFIXES` / `ENVELOPE_STANDALONE_MARKERS` that fed
  it) — channel-envelope stripper for the proactive-memory persist
  path; preserves inline bracketed content, returns `None` when nothing
  meaningful survives.
- `safe_trim_messages` — the trim-on-turn-boundary helper for working
  history, with the rescued-pinned-message reinsertion and the
  post-trim re-validate / ensure-starts-with-user safety net.
- `strip_processed_image_data`, `strip_prior_image_data` — image-data
  trimmers for already-processed turns.
- `accumulate_token_usage` — per-turn `TokenUsage` rollup.
- `sanitize_tool_result_content` — strip injection markers + dynamic
  truncate, delegating to the `ContextEngine` when present.
- `sanitize_sender_label` — bounded `[name]:` prefix sanitizer for
  group-chat senders.

mod.rs drops the now-unused `truncate_tool_result_dynamic`,
`ERR_PATH_TRAVERSAL`, `ERR_SANDBOX_ESCAPE` top-level imports — all
three were only consumed by the helpers being extracted. The big
`mod tests` block picks up `ACCUMULATED_TEXT_MAX_BYTES`,
`ENVELOPE_LINE_PREFIXES`, and `ENVELOPE_STANDALONE_MARKERS` via the
same test-only `use super::message::*` / `use crate::silent_response::*`
pattern the prior commits established.

The single stranded `/// Check if a response is a silent-reply sentinel.`
doc-line that used to sit above the moved `is_no_reply` (and which
the deletion range missed by one line) is removed in this commit too,
fixing the resulting `clippy::empty-line-after-doc-comments` lint on
the `tool_use_blocks_from_calls` boundary.

Behavior-preserving. mod.rs: 13,027 -> 12,641 LOC (-386).
cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
